# tests/openapi.yaml
openapi: 3.0.3
info:
  title: Tiikii Festival API
  description: API para el sistema de festivales Tiikii con funcionalidades de tickets, chat, mesh network y más
  version: 1.0.0
  contact:
    name: Tiikii Development Team
    email: dev@tiikii.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: https://api.tiikii.com
    description: Servidor de producción
security:
  - BearerAuth: []
paths:
...
  /api/festivals/{id}:
    get:
      summary: Obtener festival por ID
      description: Devuelve la información detallada de un festival
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Información del festival
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/FestivalDetail'
                  message:
                    type: string
                    example: "Festival retrieved successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/festivals/{id}/tickets-preview:
    get:
      summary: Vista previa de tickets del festival
      description: Obtiene una vista previa pública de los tickets disponibles para un festival
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Vista previa de tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                    example: "festival-123"
                  previewOnly:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login to see full ticket details and purchase"
                  templates:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        price:
                          type: number
                        currency:
                          type: string
                        isAvailable:
                          type: boolean
  /api/festivals/{id}/stats:
    get:
      summary: Estadísticas del festival
      description: Obtiene estadísticas detalladas de un festival
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Estadísticas del festival
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                  festivalName:
                    type: string
                  stats:
                    type: object
                    properties:
                      totalTickets:
                        type: integer
                      activeTickets:
                        type: integer
                      totalArtists:
                        type: integer
                      totalStages:
                        type: integer
                      totalVendors:
                        type: integer
                      totalPOIs:
                        type: integer
  /api/festivals/{id}/map:
    get:
      summary: Datos del mapa del festival
      description: Obtiene todos los datos relacionados con el mapa del festival
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Datos del mapa
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                  mapData:
                    type: object
                    properties:
                      stages:
                        type: array
                        items:
                          $ref: '#/components/schemas/Stage'
                      vendors:
                        type: array
                        items:
                          $ref: '#/components/schemas/Vendor'
                      pois:
                        type: array
                        items:
                          $ref: '#/components/schemas/POI'
                      geofences:
                        type: array
                        items:
                          $ref: '#/components/schemas/Geofence'
  /api/festivals/search/{query}:
    get:
      summary: Buscar festivales
      description: Busca festivales por nombre, descripción o venue
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
          example: "electronic"
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  festivals:
                    type: array
                    items:
                      $ref: '#/components/schemas/Festival'
                  count:
                    type: integer
...
  /api/artists/festival/{festivalId}:
    get:
      summary: Obtener artistas de festival
      description: Lista todos los artistas de un festival específico
      parameters:
        - name: festivalId
          in: path
          required: true
          schema:
            type: string
          example: "festival-123"
        - name: genre
          in: query
          schema:
            type: string
          example: "electronic"
      responses:
        '200':
          description: Lista de artistas del festival
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                  festivalName:
                    type: string
                  artists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artist'
                  count:
                    type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Festival:
      type: object
      properties:
        id:
          type: string
          example: "festival-123"
        name:
          type: string
          example: "Tiikii Festival 2024"
        description:
          type: string
          example: "El mejor festival de música electrónica"
        logo:
          type: string
          example: "https://example.com/logo.png"
        venue:
          type: string
          example: "Parque Central"
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
        primaryColor:
          type: string
          example: "#FF6B6B"
        secondaryColor:
          type: string
          example: "#4ECDC4"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    FestivalDetail:
      allOf:
        - $ref: '#/components/schemas/Festival'
        - type: object
          properties:
            latitudeDelta:
              type: number
              example: 0.01
            longitudeDelta:
              type: number
              example: 0.01
            accentColor:
              type: string
              example: "#45B7D1"
            backgroundColor:
              type: string
              example: "#F8F9FA"
            decorationIcons:
              type: array
              items:
                type: string
              example: ["🎵", "🎪", "🎨"]
...
    Stage:
      type: object
      properties:
        id:
          type: string
          example: "stage-123"
        festivalId:
          type: string
          example: "festival-123"
        name:
          type: string
          example: "Main Stage"
        description:
          type: string
          example: "The main stage of the festival"
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
        capacity:
          type: integer
          example: 10000
        isActive:
          type: boolean
          example: true
    Vendor:
      type: object
      properties:
        id:
          type: string
          example: "vendor-123"
        festivalId:
          type: string
          example: "festival-123"
        name:
          type: string
          example: "Food Truck Central"
        description:
          type: string
          example: "Delicious food and drinks"
        category:
          type: string
          example: "food"
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
        isActive:
          type: boolean
          example: true
    POI:
      type: object
      properties:
        id:
          type: string
          example: "poi-123"
        festivalId:
          type: string
          example: "festival-123"
        name:
          type: string
          example: "First Aid Station"
        description:
          type: string
          example: "Medical assistance and first aid"
        category:
          type: string
          example: "medical"
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
        isActive:
          type: boolean
          example: true
    Geofence:
      type: object
      properties:
        id:
          type: string
          example: "geofence-123"
        festivalId:
          type: string
          example: "festival-123"
        name:
          type: string
          example: "VIP Area"
        description:
          type: string
          example: "Exclusive VIP area"
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
        radius:
          type: number
          example: 50.0
        isActive:
          type: boolean
          example: true
...
    Artist:
      type: object
      properties:
        id:
          type: string
          example: "artist-123"
        name:
          type: string
          example: "Deadmau5"
        bio:
          type: string
          example: "Canadian electronic music producer and DJ"
        genre:
          type: string
          example: "electronic"
        imageUrl:
          type: string
          example: "https://example.com/artist.jpg"
        socialMedia:
          type: object
          example: {"twitter": "@deadmau5", "instagram": "@deadmau5"}
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            type:
              type: string
              example: "ValidationError"
            message:
              type: string
              example: "Invalid input data"
            code:
              type: string
              example: "VALIDATION_ERROR"
            field:
              type: string
              example: "email"
            details:
              type: array
              items:
                type: string
              example: ["Field 'email' is required"]
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "NotFoundError"
              message: "Resource not found"
              code: "NOT_FOUND"
...